feat : 팔로우 추가 플로우 변경
